+++ src/vfs/storage/mod.rs
@@ -7,9 +7,7 @@ mod engine;
 
 pub use engine::Config as EngineConfig;
 pub(crate) use engine::Engine;
-
 mod reader;
-mod reader_v2;
 pub(crate) mod scheduler;
 mod sto;
 mod worker;
